requirements.txt
matplotlib
PyQt5
numpy
pyqtgraph
keras
Pillow
tensorflow